% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_functional_groups.R
\name{create_functional_groups}
\alias{create_functional_groups}
\title{create_functional_groups function}
\usage{
create_functional_groups(species_list_location, species_info_csv)
}
\arguments{
\item{species_info_csv}{name of csv file with the following column headers:
\itemize{
 \item{"Genus"}
 \item{"species"}
 \item{"common name}
 \item{"Trophic level information"}{one of the following for trophic level:
 TrophicLevel (developed by user),FoodTroph (from gather_data_for_species), or
 DietTroph (from gather_data_for_species).  These will be averaged.}
 \
 \item{"parameter 2"}{Stuff}
}}

\item{species_data_location}{where csv file with species data is located (defaults to
working directory).}
}
\description{
This function uses data to create functional (focus) groups for an Atlantis model
}
\details{
This function uses provided information to suggest functional group
classification for Atlantis models.  Species are separated based on taxonomy,
depth, habitat, and trophic-level groupings.  Given the available information,
this focuses mostly on fish
}
\keyword{biology}
\keyword{prm}

