% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gather_data_for_species.R
\name{gather_data_for_species}
\alias{gather_data_for_species}
\title{gather_data_for_species function}
\usage{
gather_data_for_species(species_list_location, species_list_csv)
}
\arguments{
\item{species_list_location}{where is species list located}

\item{species_list_csv}{name of csv file with the following column headers:
Genus, species,common_name}
}
\description{
This function aids in gathering the species needed to create functional groups
and prm files in Atlantis
}
\details{
This function aids in gathering the species needed to create functional
groups and prm files in Atlantis. User submits a list of potential species to
 be included in the model. The code gathers data on species from cloud-based,
 public data sources and offers potential functional groupings.
}
\keyword{biology}
\keyword{functional}
\keyword{groups,}
\keyword{prm,}

