% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gather_group_for_species.R
\name{gather_group_for_species}
\alias{gather_group_for_species}
\title{gather_group_for_species function}
\usage{
gather_group_for_species(species_list_location = getwd(), species_list_csv)
}
\arguments{
\item{species_list_location}{where species list is located; defaults to working
directory}

\item{species_list_csv}{name of csv file with the following column headers:
Genus, species,common_name}
}
\description{
This function aids in determining the functional group and higher order classification
information for species to be included in the Atlantis model
}
\details{
This function aids in gathering higher taxonomic information and Atlantis
group types for species to be included in an Atlantis model. User submits a list of potential species to
 be included in the model. The code gathers data on species from cloud-based,
 public data sources.
}
\keyword{biology}
\keyword{functional}
\keyword{groups,}
\keyword{prm,}

