% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gather_habitat_for_species.R
\name{gather_habitat_for_species}
\alias{gather_habitat_for_species}
\title{gather_habitat_for_species function}
\usage{
gather_habitat_for_species(species_list_location = getwd(), species_list_csv,
  map_location = getwd(), map_name = "map_for_bgmeriser",
  habitat_list = c("Intertidal", "Sublittoral", "Caves", "Oceanic",
  "Epipelagic", "Mesopelagic", "Bathypelagic", "Abyssopelagic", "Hadopelagic",
  "Estuaries", "Mangroves", "MarshesSwamps", "Stream", "Lakes", "Benthos",
  "Sessile", "Demersal", "Pelagic", "Endofauna", "Megabenthos", "Meiobenthos",
  "SoftBottom", "Sand", "Coarse", "Fine", "Level", "Sloping", "Silt", "Mud",
  "Ooze", "HardBottom", "Rocky", "Rubble", "SeaGrassBeds", "BedsBivalve",
  "BedsRock", "CoralReefs", "DropOffs", "ReefFlats", "Lagoons",
  "DeepWaterCorals"))
}
\arguments{
\item{species_list_location}{where is species list located; defaults to working
director}

\item{species_list_csv}{name of csv file with the following column headers:
Genus, species,common_name.  functional_group may also be included.}

\item{map_location}{location of shape file used to create bgm; defaults to working
directory}

\item{map_name}{name of map used for bgm creation; can be produced by rbgmeriser
function or created manually; defaults to map_for_bgmeriser (produced by
rbgmeriser function)}

\item{habitat_list}{(optional)list of potential habitat types to be included in the
model. defaults to full list avaialble from Fishbase}
}
\description{
This function aids in gathering data on species occurence that may be useful
in constructing functional groups and diet matrices
}
\details{
This function aids in gathering data on species occurrence that may
be useful in constructing functional groups and diet matrices.  User submits
a list of potential species to be included in the model (may include potential
functional groupings). The code gathers data on species from rgbif and rfishbase.
}
\keyword{biology}
\keyword{functional}
\keyword{groups,}
\keyword{prm,}

