% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gather_interactions_for_species.R
\name{gather_interactions_for_species}
\alias{gather_interactions_for_species}
\title{gather_interactions_for_species function}
\usage{
gather_interactions_for_species(species_list_location = getwd(),
  species_list_csv, map_location = getwd(), map_name = "map_for_bgmeriser",
  bbox_coordinates)
}
\arguments{
\item{species_list_location}{where is species list located,defaults to working directory}

\item{species_list_csv}{name of csv file with the following column headers:
Genus, species,common_name.  functional_group may also be included.}

\item{map_location}{location of shape file used to create bgm, defaults to working
directory}

\item{map_name}{name of map used for bgm creation; can be produced by rbgmeriser
function or created manually, defaults to file created by rbgmeriser}

\item{bbox_coordinates}{(optional) bounding box (in EPSG:4326 decimal degrees, defining
"left, bottom, right, top" of bounding box) in which you wish for information
on interactions.  If these are not supplied code uses provided map to determine
bounding boxes}
}
\description{
This function aids in gathering data on species interactions that may be useful
in constructing functional groups and diet matrices
}
\details{
This function aids in gathering data on species interactions that may
be useful in constructing functional groups and diet matrices.  User submits
a list of potential species to be included in the model (may include potential
functional groupings). The code gathers data on species from rglobi.
}
\keyword{biology}
\keyword{functional}
\keyword{groups,}
\keyword{prm,}

