% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_group_parameters.R
\name{generate_group_parameters}
\alias{generate_group_parameters}
\title{generate_group_parameters function}
\usage{
generate_group_parameters(species_data_location = getwd(),
  species_info_groups_csv)
}
\arguments{
\item{species_data_location}{where csv file with species data is located (defaults to
working directory).}

\item{species_info_groups_csv}{name of csv file with the following column headers
(all provided by merging "gather" functions (gather_groups_for_species, gather_data_for_species,
gather_habitat_for_species):
\itemize{
 \item{Genus}
 \item{species}
 \item{family}
 \item{order}
 \item{class}
 \item{superclass}
 \item{phylum}
 \item{kingdom}
 \item{TL_final}
 \item {Data on depth associations, specifically
 \itemize{
 \item {min_depth}
 \item {max_depth}
 }
 }
 \item {Growth information, specifically
 \itemize{
 \item {mean_a}
 \item {mean_b}
 \item {mean_to}
 \item {mean_K}
 \item {mean_M}
 \item {max_length}
 \item {mean_Loo}
 \item {max_age}
 }
 }
 \item {atlantis_type} {classification for Atlantis group type.  For Vertebrates, these can be
 \itemize{
 \item{fish}
 \item{bird}
 \item{mammal}
 \item{shark}
 }
 for invertebrates, see a long list at https://wiki.csiro.au/display/Atlantis/AtlantisGroupTypes
 }
 \item {group_name} {a name for the specific functional or focus group each organism
 belongs in; these are suggested by create_functional_groups but may need to be
 modified for readabilty (pick names you can remember and choose between!)}
 \item {group_code} {a 2 or 3-letter abbreviation for group_name}
 }}
}
\description{
This function calculates the group values needed for Atlantis
}
\details{
This function uses provided information to Atlantis parameters for each
functional group
}
\keyword{biology}
\keyword{prm}

