% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper_functions_ratlantis.R
\name{\%!in\%}
\alias{\%!in\%}
\title{"Not in" function}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{}

\item{table}{}
}
\description{
This function allows you to check if an element is not in a larger group
}

