% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_functions_ratlantis.R
\name{maxnona}
\alias{maxnona}
\title{maxnona function}
\usage{
maxnona(...)
}
\arguments{
\item{...}{list for which you seeking a maximum}
}
\description{
This function allows you to calculate the max value of a set while excluuding
NA's
}

