% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper_functions_ratlantis.R
\name{meannona}
\alias{meannona}
\title{meannona function}
\usage{
meannona(...)
}
\arguments{
\item{...}{list for which you seeking a mean}
}
\description{
This function allows you to calculate the mean value of a set while excluuding
NA's
}

