% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper_functions_ratlantis.R
\name{minnona}
\alias{minnona}
\title{minnona function}
\usage{
minnona(...)
}
\arguments{
\item{...}{list for which you seeking a minimum}
}
\description{
This function allows you to calculate the mean value of a set while excluuding
NA's
}

