% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper_functions_ratlantis.R
\name{not_in}
\alias{not_in}
\title{"not in" function}
\usage{
not_in(x, table)
}
\arguments{
\item{x}{object you are trying to check for}

\item{table}{group you are checking against}
}
\description{
This function allows you to check if an element is not in a larger group
}

