% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helper_functions_ratlantis.R
\name{points2polygons}
\alias{points2polygons}
\title{Points to Polygon function}
\usage{
points2polygons(df, data)
}
\arguments{
\item{df}{}

\item{data}{}
}
\description{
This function allows you to change points to polygons
}

