% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ratlantis-package.r
\docType{package}
\name{ratlantis}
\alias{ratlantis}
\alias{ratlantis-package}
\title{ratlantis.}
\description{
The ratlantis package provides a way to develop Atlantis
models in the R environment.
}

