% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rbgmeriser.R
\name{rbgmeriser}
\alias{rbgmeriser}
\title{rbgmeriser function}
\usage{
rbgmeriser(map_location, map_name, boundary_boxes = NULL,
  get_bathymetry = TRUE, bathymetry_layer_location, bathymetry_layer_name,
  bathymetry_cutoff = 0.9, bathymetry_levels = c(-10, -20, -50, -200, -1000,
  -2000, -4000), bgmeriser_location)
}
\arguments{
\item{map_location}{location of gis layer stored in wgs84 format}

\item{map_name}{name of map to be used for bgm creation. Must incude "box_id"
column. Other columns may be included (horizmix, vertmix,boundary_code)}

\item{boundary_boxes}{a list of boxes that boundary boxes in the model}

\item{get_bathymetry}{Should function automatically add depths (T(rue) or
F(alse)}

\item{bathymetry_layer_location}{location of gis layer stored in wgs84 format.}

\item{bathymetry_layer_name}{name of bathmetry shp layer. Must be in wgs84
format}

\item{bathymetry_cutoff}{What quantile should be assigned to polygon. Defaults
to 90th quantile (to avoid small number of canyons or slope biasing entire
polygon)}

\item{bathymetry_levels}{Depth layers for final Atlantis polygons}

\item{bgmeriser_location}{location of bgmeriser.jar file (location also must
include java.exe file and may not have spaces (e.g. not C:/Desktop/John Doe/bgm))}
}
\description{
This function creates bgm file format needed for Atlantis
}
\details{
This function creates bgm file format needed for Atlantis by calling
java program. It also adds in depth data (if desired) using provided
bathymetry layer.  This functions as a wrapper for bgmeriser.jar, available
from CSIRO, and adds in required information.
}
\keyword{bgm}

