% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bgmeriser.R
\name{rbgmerizer}
\alias{rbgmerizer}
\title{bgmerizer function}
\usage{
rbgmerizer(map_location = getwd(), map_name, boundary_boxes,
  bgmerizer_location = getwd())
}
\arguments{
\item{map_location}{location of gis layer stored in wgs84 format.  Defaults
to working directory}

\item{map_name}{name of map to be used for bgm creation. Must incude box_id
column and boundary column. Other columns may be included}

\item{boundary_boxes}{a list of boxes that boundary boxes in the model}

\item{bgmerizer_location}{location of bgmerizer.jar file, assumed to be in
working directory along with java.exe file}
}
\description{
This function creates bgm file format needed for Atlantis
}
\details{
This function creates bgm file format needed for Atlantis by calling
java program. It also adds in depth data (if desired) using provided
bathymetry layer.
layer.
}
\keyword{bgm}

