% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_biology_prm.R
\name{create_biology_prm}
\alias{create_biology_prm}
\title{create_biology_prm function}
\usage{
create_biology_prm(species_data_location = getwd(), group_data_csv)
}
\arguments{
\item{species_data_location}{where csv file with species data is located (defaults to
working directory).}

\item{species_info_groups_csv}{name of csv file that must contain the following column headers
(all provided by merging generate_group_parameter function):
\itemize{
 \item{atlantis_type}
 \item{group_name}
 \item{group_code}
 \item{TL_final}
 \item{mean_M}
 \item{max_age}
 }
 The following columns are optional (will be filled in by function if not
 provided)
 \itemize{
 \item{num_of_age_classes}
 }}
}
\description{
This function creates the biology prm file needed for Atlantis
}
\details{
This function creates the biology prm file needed by Atlantis.
}
\keyword{biology}
\keyword{prm}

