% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_biology_prm.R
\name{create_biology_prm}
\alias{create_biology_prm}
\title{create_biology_prm function}
\usage{
create_biology_prm(species_file)
}
\arguments{
\item{species_file}{}
}
\description{
This function creates the biology prm file needed for Atlantis
}
\details{
This function creates the biology prm file needed by Atlantis. User
create functional groupings using the create_functional_groups function or supply
their own list with needed parameters.
}
\keyword{biology}
\keyword{prm}

