% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_biology_prm.R
\name{create_biology_prm}
\alias{create_biology_prm}
\title{create_biology_prm function}
\usage{
create_biology_prm()
}
\arguments{
\item{map_name}{name of map to be used for bgm creation. Must incude "box_id"
column. Other columns may be included (horizmix, vertmix,boundary_code)}
}
\description{
This function creates the biology prm file needed for Atlantis
}
\details{
This function creates the biology prm file needed by Atlantis. User
create functional groupings using the create_functional_groups function or supply
their own list with needed parameters.
}
\keyword{biology}
\keyword{prm}

