% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_functional_groups.R
\name{create_functional_groups}
\alias{create_functional_groups}
\title{create_functional_groups function}
\usage{
create_functional_groups(species_list_location, species_info_csv)
}
\arguments{
\item{species_list_location}{where species list is located}

\item{species_info_csv}{name of csv file with the following column headers:
Genus, species,common_name}
}
\description{
This function creates the biology prm file needed for Atlantis
}
\details{
This function uses provided information to suggest functional group
classification for Atlantis models
}
\keyword{biology}
\keyword{prm}

