% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_functional_groups.R
\name{create_functional_groups}
\alias{create_functional_groups}
\title{create_functional_groups function}
\usage{
create_functional_groups()
}
\arguments{
\item{map_location}{location of gis layer stored in wgs84 format}
}
\description{
This function creates the biology prm file needed for Atlantis
}
\details{
This function creates the biology prm file needed by Atlantis. User
submits a list of potential species to be included in the model. The code gathers
data on species from cloud-based, public data sources and offers potential functional
groupings.
}
\keyword{biology}
\keyword{prm}

