% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gather_habitat_for_species.R
\name{gather_habitat_for_species}
\alias{gather_habitat_for_species}
\title{gather_habitat_for_species function}
\usage{
gather_habitat_for_species(species_list_location, species_list_csv,
  map_location = "empty", map_name = "empty",
  habitat_list = c("estuaries", "mangrove", "intertidal", "soft", "rocky",
  "marine", "oceanic", "neritic", "coral reefs", "soft bottom", "hard bottom",
  "seagrass beds", "reef", "sand", "benthic", "slope"))
}
\arguments{
\item{species_list_location}{where is species list located}

\item{species_list_csv}{name of csv file with the following column headers:
Genus, species,common_name.  functional_group may also be included.}

\item{map_location}{location of shape file used to create bgm}

\item{map_name}{name of map used for bgm creation; can be produced by rbgmeriser
function or created manually}

\item{habitat_list}{list of potential habitat types to be included in the
model}
}
\description{
This function aids in gathering data on species occurence that may be useful
in constructing functional groups and diet matrices
}
\details{
This function aids in gathering data on species occurrence that may
be useful in constructing functional groups and diet matrices.  User submits
a list of potential species to be included in the model (may include potential
functional groupings). The code gathers data on species from rgbif and rfishbase.
}
\keyword{biology}
\keyword{functional}
\keyword{groups,}
\keyword{prm,}

